some error handling added

+++ extensions/lab.py
@@ -2,7 +2,6 @@
 import disnake
 from disnake.ext import commands, tasks
 import datetime
-import time
 from bs4 import BeautifulSoup
 import asyncio
 from config import BOT
@@ -12,18 +11,19 @@ class Lab(commands.Cog):
     def __init__(self, bot):
         self.bot: commands.Bot = bot
         self.task_update_labs.start()
+        self.lab_list = ["UBER LAB", "MERC LAB", "CRUEL LAB", "NORMAL LAB"]
+        self.last_full_lab = disnake.utils.utcnow() - datetime.timedelta(days=9)
 
     async def get_labs(self) -> dict:
-        lab_list = ["UBER LAB", "MERC LAB", "CRUEL LAB", "NORMAL LAB"]
         async with httpx.AsyncClient(follow_redirects=True, timeout=10) as c:
             r = await c.get("https://www.poelab.com/")
             soup = BeautifulSoup(r, "html5lib")
-            labs = {lab: soup.find("a", string=lab)["href"] for lab in lab_list}
+            labs = {lab: soup.find("a", string=lab)["href"] for lab in self.lab_list}
             assert len(labs) == 4
             print(labs)
             updated_labs = {}
             async with asyncio.TaskGroup() as tg:
-                for lab in lab_list:
+                for lab in self.lab_list:
                     updated_labs[lab] = tg.create_task(self.get_lab(labs[lab], c))
             all_labs = {x: y.result() for x, y in updated_labs.items()}
             print(all_labs)
@@ -45,32 +45,55 @@ class Lab(commands.Cog):
         await self._update_labs()
         await inter.edit_original_message("done!")
 
-    @tasks.loop(time=[datetime.time(n, 30) for n in range(24)])
+    @tasks.loop(time=[datetime.time(h, m) for h in range(24) for m in range(0, 60, 10)])
     async def task_update_labs(self):
         await self._update_labs()
 
     async def _update_labs(self):
+        now = disnake.utils.utcnow()
+        if self.last_full_lab.date() == now.date():
+            return
+
+        try:
+            all_labs = await self.get_labs()
+        except:
+            return
+
+        up_to_date_counter = len(
+            [
+                True
+                for lab in self.lab_list
+                if now.strftime(r"%B %d, %Y") == all_labs[lab][1]
+            ]
+        )
+        if up_to_date_counter == 0:
+            return
+
         channel = self.bot.get_channel(BOT["LAB_CHANNEL"])
-        async for message in channel.history(limit=10):
+        async for message in channel.history(limit=20):
             await message.delete()
             await asyncio.sleep(0.2)
-        all_labs = await self.get_labs()
-        for lab in reversed(["UBER LAB", "MERC LAB", "CRUEL LAB", "NORMAL LAB"]):
+        for lab in reversed(self.lab_list):
             embed = disnake.Embed(
                 title=lab, description=f"Data: {all_labs[lab][1]}", url=all_labs[lab][0]
             )
             embed.set_image(url=all_labs[lab][2])
             await channel.send(embed=embed)
             await asyncio.sleep(0.2)
-        now = datetime.datetime.utcnow()
-        next_midnight = now.replace(hour=0, minute=0, second=0, microsecond=0)
-        if now > next_midnight:
-            next_midnight += datetime.timedelta(days=1)
-        next_midnight -= datetime.timedelta(hours=3)
-        unix_time = int(time.mktime(next_midnight.timetuple()))
+
+        next_midnight = now.replace(
+            hour=0, minute=0, second=0, microsecond=0
+        ) + datetime.timedelta(days=1)
         await channel.send(
-            f"Próximo lab: <t:{unix_time}:R>\nSe a data acima já passou, as imagens ainda estão desatualizadas."
+            f"Próximo lab: {disnake.utils.format_dt(next_midnight, style='R')}\nSe a data acima já passou, as imagens ainda estão desatualizadas."
         )
+        if up_to_date_counter < 4:
+            singular = up_to_date_counter == 1
+            await channel.send(
+                f"Atenção: Apenas {up_to_date_counter} dos {len(self.lab_list)} labs {'foi' if singular else 'foram'} atualizado{'' if singular else 's'} pelo poelab.com até agora."
+            )
+        else:
+            self.last_full_lab = disnake.utils.utcnow()
 
 
 def setup(bot):
